# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BulkImportResult": {
        "additionalProperties": false,
        "properties": {
          "created_ids": {
            "items": {
              "type": "string"
            },
            "title": "Created Ids",
            "type": "array"
          },
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ImportErrorDetail"
            },
            "title": "Errors",
            "type": "array"
          },
          "failed_rows": {
            "title": "Failed Rows",
            "type": "integer"
          },
          "skipped_rows": {
            "default": 0,
            "title": "Skipped Rows",
            "type": "integer"
          },
          "success_rows": {
            "title": "Success Rows",
            "type": "integer"
          },
          "total_rows": {
            "title": "Total Rows",
            "type": "integer"
          }
        },
        "required": [
          "total_rows",
          "success_rows",
          "failed_rows"
        ],
        "title": "BulkImportResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "row": {
            "title": "Row",
            "type": "integer"
          }
        },
        "required": [
          "row",
          "field",
          "reason"
        ],
        "title": "ImportErrorDetail",
        "type": "object"
      },
...
      "KnowledgeBulkImportRequest": {
        "additionalProperties": false,
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeImportRow"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "KnowledgeBulkImportRequest",
        "type": "object"
      },
      "KnowledgeEntryCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "notes": {
            "default": "",
            "maxLength": 8000,
            "title": "Notes",
            "type": "string"
          },
          "problem": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Problem",
            "type": "string"
          },
          "related_item_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Item Ids",
            "type": "array"
          },
          "root_cause": {
            "default": "",
            "maxLength": 8000,
            "title": "Root Cause",
            "type": "string"
          },
          "solution": {
            "maxLength": 12000,
            "minLength": 1,
            "title": "Solution",
            "type": "string"
          },
          "source_ref": {
            "default": "",
            "maxLength": 2000,
            "title": "Source Ref",
            "type": "string"
          },
          "source_type": {
            "default": "manual",
            "enum": [
              "manual",
              "document-derived",
              "autotest-derived"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "reviewed",
              "verified",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "default": "",
            "maxLength": 2000,
            "title": "Tags",
            "type": "string"
          },
          "title": {
            "default": "",
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "problem",
          "solution"
        ],
        "title": "KnowledgeEntryCreateRequest",
        "type": "object"
      },
      "KnowledgeEntryPageResponse": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeEntryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "KnowledgeEntryPageResponse",
        "type": "object"
      },
      "KnowledgeEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "problem": {
            "title": "Problem",
            "type": "string"
          },
          "related_item_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Item Ids",
            "type": "array"
          },
          "root_cause": {
            "title": "Root Cause",
            "type": "string"
          },
          "solution": {
            "title": "Solution",
            "type": "string"
          },
          "source_ref": {
            "default": "",
            "title": "Source Ref",
            "type": "string"
          },
          "source_type": {
            "default": "manual",
            "enum": [
              "manual",
              "document-derived",
              "autotest-derived"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "reviewed",
              "verified",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "title": "Tags",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "problem",
          "root_cause",
          "solution",
          "tags",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeEntryResponse",
        "type": "object"
      },
      "KnowledgeEntryUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "change_note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Note"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "problem": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem"
          },
          "related_item_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Item Ids"
          },
          "root_cause": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause"
          },
          "solution": {
            "anyOf": [
              {
                "maxLength": 12000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solution"
          },
          "source_ref": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "source_type": {
            "anyOf": [
              {
                "enum": [
                  "manual",
                  "document-derived",
                  "autotest-derived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "draft",
                  "reviewed",
                  "verified",
                  "archived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tags": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "KnowledgeEntryUpdateRequest",
        "type": "object"
      },
      "KnowledgeImportRow": {
        "additionalProperties": false,
        "properties": {
          "row_number": {
            "minimum": 1.0,
            "title": "Row Number",
            "type": "integer"
          },
          "values": {
            "$ref": "#/components/schemas/KnowledgeEntryCreateRequest"
          }
        },
        "required": [
          "row_number",
          "values"
        ],
        "title": "KnowledgeImportRow",
        "type": "object"
      },
      "KnowledgeRevisionDiffItem": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "new_value": {
            "title": "New Value",
            "type": "string"
          },
          "old_value": {
            "title": "Old Value",
            "type": "string"
          }
        },
        "required": [
          "field",
          "old_value",
          "new_value"
        ],
        "title": "KnowledgeRevisionDiffItem",
        "type": "object"
      },
      "KnowledgeRevisionDiffResponse": {
        "additionalProperties": false,
        "properties": {
          "changed": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeRevisionDiffItem"
            },
            "title": "Changed",
            "type": "array"
          },
          "entry_id": {
            "title": "Entry Id",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          }
        },
        "required": [
          "revision_id",
          "entry_id"
        ],
        "title": "KnowledgeRevisionDiffResponse",
        "type": "object"
      },
      "KnowledgeRevisionResponse": {
        "additionalProperties": false,
        "properties": {
          "change_note": {
            "title": "Change Note",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "entry_id": {
            "title": "Entry Id",
            "type": "string"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "problem": {
            "title": "Problem",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "root_cause": {
            "title": "Root Cause",
            "type": "string"
          },
          "solution": {
            "title": "Solution",
            "type": "string"
          },
          "source_ref": {
            "default": "",
            "title": "Source Ref",
            "type": "string"
          },
          "source_type": {
            "default": "manual",
            "enum": [
              "manual",
              "document-derived",
              "autotest-derived"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "reviewed",
              "verified",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "title": "Tags",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "revision_id",
          "entry_id",
          "version_number",
          "title",
          "problem",
          "root_cause",
          "solution",
          "tags",
          "notes",
          "change_note",
          "created_at"
        ],
        "title": "KnowledgeRevisionResponse",
        "type": "object"
      },
      "LogbookBulkImportRequest": {
        "additionalProperties": false,
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/LogbookImportRow"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "LogbookBulkImportRequest",
        "type": "object"
      },
...
      "LogbookImportRow": {
        "additionalProperties": false,
        "properties": {
          "row_number": {
            "minimum": 1.0,
            "title": "Row Number",
            "type": "integer"
          },
          "values": {
            "$ref": "#/components/schemas/LogbookEntryCreateRequest"
          }
        },
        "required": [
          "row_number",
          "values"
        ],
        "title": "LogbookImportRow",
        "type": "object"
      },
...
      "MessageResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "PromptBulkImportRequest": {
        "additionalProperties": false,
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PromptImportRow"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "PromptBulkImportRequest",
        "type": "object"
      },
      "PromptImportRow": {
        "additionalProperties": false,
        "properties": {
          "row_number": {
            "minimum": 1.0,
            "title": "Row Number",
            "type": "integer"
          },
          "values": {
            "$ref": "#/components/schemas/SavedPromptCreateRequest"
          }
        },
        "required": [
          "row_number",
          "values"
        ],
        "title": "PromptImportRow",
        "type": "object"
      },
...
      "SavedPromptCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 12000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "tags": {
            "default": "",
            "maxLength": 2000,
            "title": "Tags",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "SavedPromptCreateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Knowledge Workspace API",
    "version": "5.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/import/knowledge": {
      "post": {
        "operationId": "import_knowledge_entries_api_import_knowledge_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBulkImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Knowledge Entries"
      }
    },
    "/api/import/logbook": {
      "post": {
        "operationId": "import_logbook_entries_api_import_logbook_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogbookBulkImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Logbook Entries"
      }
    },
    "/api/import/prompts": {
      "post": {
        "operationId": "import_prompts_api_import_prompts_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptBulkImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Prompts"
      }
    },
...
    "/api/knowledge/entries": {
      "get": {
        "operationId": "list_knowledge_entries_api_knowledge_entries_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntryPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Knowledge Entries"
      },
      "post": {
        "operationId": "create_knowledge_entry_api_knowledge_entries_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeEntryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Knowledge Entry"
      }
    },
    "/api/knowledge/entries/{entry_id}": {
      "patch": {
        "operationId": "update_knowledge_entry_api_knowledge_entries__entry_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeEntryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Knowledge Entry"
      }
    },
    "/api/knowledge/{entry_id}/revisions": {
      "get": {
        "operationId": "list_knowledge_revisions_api_knowledge__entry_id__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeRevisionResponse"
                  },
                  "title": "Response List Knowledge Revisions Api Knowledge  Entry Id  Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Knowledge Revisions"
      }
    },
    "/api/knowledge/{entry_id}/revisions/{revision_id}/diff": {
      "get": {
        "operationId": "get_knowledge_revision_diff_api_knowledge__entry_id__revisions__revision_id__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeRevisionDiffResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Knowledge Revision Diff"
      }
    },
    "/api/knowledge/{entry_id}/revisions/{revision_id}/restore": {
      "post": {
        "operationId": "restore_knowledge_revision_api_knowledge__entry_id__revisions__revision_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Knowledge Revision"
      }
    },
...
  }
}